+++ src/brickingbad/domain/game/persistence/Save.java
@@ -5,14 +5,14 @@ public class Save {
   private String name;
   private int score;
   private int lives;
-  private int paddleX;
-  private int paddleY;
-//  private int[] brickX;
-//  private int[] brickY;
-//  private int[] ballX;
-//  private int[] ballY;
-//  private int[] ballVelX;
-//  private int[] ballVelY;
+  private double paddleX;
+  private double paddleY;
+//  private double[] brickX;
+//  private double[] brickY;
+//  private double[] ballX;
+//  private double[] ballY;
+//  private double[] ballVelX;
+//  private double[] ballVelY;
 
   public int getScore() {
     return score;
@@ -30,67 +30,67 @@ public class Save {
     this.lives = lives;
   }
 
-//  public int[] getBallX() {
+//  public double[] getBallX() {
 //    return ballX;
 //  }
 //
-//  public void setBallX(int[] ballX) {
+//  public void setBallX(double[] ballX) {
 //    this.ballX = ballX;
 //  }
 //
-//  public int[] getBallY() {
+//  public double[] getBallY() {
 //    return ballY;
 //  }
 //
-//  public void setBallY(int[] ballY) {
+//  public void setBallY(double[] ballY) {
 //    this.ballY = ballY;
 //  }
 //
-//  public int[] getBallVelX() {
+//  public double[] getBallVelX() {
 //    return ballVelX;
 //  }
 //
-//  public void setBallVelX(int[] ballVelX) {
+//  public void setBallVelX(double[] ballVelX) {
 //    this.ballVelX = ballVelX;
 //  }
 //
-//  public int[] getBallVelY() {
+//  public double[] getBallVelY() {
 //    return ballVelY;
 //  }
 //
-//  public void setBallVelY(int[] ballVelY) {
+//  public void setBallVelY(v[] ballVelY) {
 //    this.ballVelY = ballVelY;
 //  }
 
-  public int getPaddleX() {
+  public double getPaddleX() {
     return paddleX;
   }
 
-  public void setPaddleX(int paddleX) {
+  public void setPaddleX(double paddleX) {
     this.paddleX = paddleX;
   }
 
-  public int getPaddleY() {
+  public double getPaddleY() {
     return paddleY;
   }
 
-  public void setPaddleY(int paddleY) {
+  public void setPaddleY(double paddleY) {
     this.paddleY = paddleY;
   }
 
-//  public int[] getBrickX() {
+//  public double[] getBrickX() {
 //    return brickX;
 //  }
 //
-//  public void setBrickX(int[] brickX) {
+//  public void setBrickX(double[] brickX) {
 //    this.brickX = brickX;
 //  }
 //
-//  public int[] getBrickY() {
+//  public double[] getBrickY() {
 //    return brickY;
 //  }
 //
-//  public void setBrickY(int[] brickY) {
+//  public void setBrickY(double[] brickY) {
 //    this.brickY = brickY;
 //  }
